sort: move heap to different file
